Reddit_api/requirements.txt
praw==7.8.1
pymongo==4.8.0
python-dotenv==1.0.1
emoji==2.14.0
langdetect==1.0.9
deep-translator==1.11.4
Unidecode==1.3.8
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
numpy==1.24.3
plotly==5.15.0
spacy==3.7.2
requests==2.31.0
tqdm==4.66.1
transformers==4.35.2
torch==2.1.0
scipy==1.11.4
flask==3.0.0
